requirements.txt
django
djangorestframework
mysqlclient
requests
uvicorn
uvicorn-worker
azure-storage-blob
python-dotenv